+++ src/core/index.ts
@@ -1,4 +1,5 @@
 export * from "./auto-generator";
+export * from "./global-statistics";
 export * from "./note-generator";
 export * from "./period-index";
 export * from "./periodic-note-indexer";