+++ src/main.ts
@@ -8,6 +8,7 @@ import { PeriodChildrenBasesModal } from "./components/period-children/bases-mod
 import { TimeBudgetBlockRenderer } from "./components/time-budget";
 import type { PeriodType } from "./constants";
 import { AutoGenerator, formatAutoGenerationSummary } from "./core/auto-generator";
+import { GlobalStatisticsAggregator } from "./core/global-statistics";
 import { PeriodIndex } from "./core/period-index";
 import { PeriodicNoteIndexer } from "./core/periodic-note-indexer";
 import { SettingsStore } from "./core/settings-store";
@@ -37,6 +38,7 @@ export default class PeriodicPlannerPlugin extends Plugin {
 	autoGenerator!: AutoGenerator;
 	indexer!: PeriodicNoteIndexer;
 	periodIndex!: PeriodIndex;
+	globalStatsAggregator!: GlobalStatisticsAggregator;
 	templateService!: TemplateService;
 	private ribbonIconEl: HTMLElement | null = null;
 
@@ -61,6 +63,7 @@ export default class PeriodicPlannerPlugin extends Plugin {
 
 	onunload() {
 		this.autoGenerator.destroy();
+		this.globalStatsAggregator.destroy();
 		this.indexer.stop();
 		this.periodIndex.destroy();
 		this.templateService.destroy();
@@ -111,6 +114,7 @@ export default class PeriodicPlannerPlugin extends Plugin {
 		this.templateService = new TemplateService(this.app, this.settingsStore.settings$);
 		this.autoGenerator = new AutoGenerator(this.app, this.settingsStore.settings$, this.templateService);
 		this.periodIndex = new PeriodIndex(this.indexer);
+		this.globalStatsAggregator = new GlobalStatisticsAggregator(this.periodIndex, this.settingsStore.currentSettings);
 
 		await this.indexer.start();